LayoutTests/requestidlecallback/requestidlecallback-deadline-shortened-by-rendering-update-expected.txt

This tests that the deadline can be reached.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS idleDeadline.timeRemaining() > preferredRenderingUpdateInterval is true
PASS idleDeadline.timeRemaining() <= preferredRenderingUpdateInterval + 0.5 is true
PASS didRunIdleCallback is true
PASS successfullyParsed is true

TEST COMPLETE
hello
